GitArena/tests/requirements.txt
pytest
selenium
webdriver-manager
pytest-html
pytest-playwright
playwright
PyJWT
pytest-base-url
